# Conflicts:
#	inc/app/image.hpp
#	inc/app/scene.hpp
#	inc/lights/light_source.hpp
#	inc/lights/point_light.hpp
#	src/app/image.cpp
#	src/app/scene.cpp
#	src/lights/point_light.cpp
#	src/objects/plane.cpp
#	src/objects/sphere.cpp
#	src/objects/triangle.cpp
#	src/ray.cpp

+++ src/objects/plane.cpp
@@ -4,7 +4,7 @@
 
 #include "objects/plane.hpp"
 
-constexpr double kEpsilon = 0.0001f;
+constexpr float kEpsilon = 0.0001f;
 
 bool
 Plane::testIntersections(const Ray &cast_ray, glm::vec3 &int_point, glm::vec3 &loc_normal) const {
